roadkeeper/requirements.txt
pandas
mysql-connector-python
streamlit
folium
streamlit-folium
streamlit-webrtc
ultralytics
opencv-python
